solve circular dependecy
