+++ editor/src/utils/deep-equality-instances.ts
@@ -21,7 +21,6 @@ import type { HigherOrderControl } from '../components/canvas/canvas-types'
 import type { JSXElementName } from '../core/shared/element-template'
 import type {
   ElementPath,
-  ElementPropertyPath,
   PropertyPath,
   StaticElementPath,
 } from '../core/shared/project-file-types'
@@ -41,7 +40,6 @@ import {
 import type { LayoutTargetableProp } from '../core/layout/layout-helpers-new'
 import type { CanvasPoint, WindowPoint } from '../core/shared/math-utils'
 import { canvasPoint, windowPoint } from '../core/shared/math-utils'
-import * as EPP from '../components/template-property-path'
 
 export const ElementPathKeepDeepEquality: KeepDeepEqualityCall<ElementPath> =
   createCallFromEqualsFunction((oldPath: ElementPath, newPath: ElementPath) => {
@@ -60,16 +58,6 @@ export function PropertyPathKeepDeepEquality(): KeepDeepEqualityCall<PropertyPat
   return createCallFromEqualsFunction((oldPath: PropertyPath, newPath: PropertyPath) => {
     return PP.pathsEqual(oldPath, newPath)
   })
-}
-
-export function ElementPropertyPathKeepDeepEquality(): KeepDeepEqualityCall<ElementPropertyPath> {
-  return combine2EqualityCalls(
-    (e) => e.elementPath,
-    ElementPathKeepDeepEquality,
-    (e) => e.propertyPath,
-    PropertyPathKeepDeepEquality(),
-    (e, p) => EPP.create(e, p),
-  )
 }
 
 export function HigherOrderControlKeepDeepEquality(
